Creation of our first views.
We've generated some dummy cities with real data from our API explorer. Then, we created a ListView to show the names of those cities, and a new activity that receives those names as a string, whose ID we've decided to put in a separate class, along with the rest of the IDs of the app.