Change from ArrayAdapter to CursorAdapter.
We've changed our adapter once again to implement a CursorAdapter, erasing the old methods and implementing the new. In the "newView()" method, we'll only inflate the view, because in the "bindView()" method, we're getting a new City based on the cursor to set the data. To do so, we had to create a new constructor on our "City" model that would initialize every attribute based on the cursor given. Then, in our common class, we've created a method that will insert new cities in our DAO (it used to be only if it was the first time running the app, but we found a bug there that we'll have to solve later). In the "CityActivity", we've created an AsyncTask class that will fetch the data from our API in background and will display it when finished (controlling the back button also). The changes made in files "WeatherDbHelper", "WeatherDbContract" and "WeatherDAO" are for personal convenience.